requirements.txt
about-time==4.2.1
alive-progress==3.0.1
grapheme==0.6.0
ImageHash==4.3.1
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
PyWavelets==1.4.1
scipy==1.10.0
